requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1

#$env:DISCORD_TOKEN="SEU_TOKEN_AQUI"
#$env:BOT_PREFIX="!"

